extract out FullScreenSwipeToDismissListener
